+++ extensions/google-chrome/src/new-window.tsx
@@ -1,0 +1,10 @@
+import { showHUD } from "@raycast/api";
+import { createNewWindow } from "./actions";
+
+export default async function Command() {
+  try {
+    await createNewWindow();
+  } catch {
+    await showHUD("❌ Failed opening a new Google Chrome window");
+  }
+}