+++ extensions/google-chrome/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Google Chrome Changelog
 
+## [New Window] - 2023-03-22
+
+- Add the ability to create new window
+
 ## [Tab closing] - 2023-03-10
 
 - Add the ability to close tabs